+++ tests/unit/EnrichedCustomerDataTest.php
@@ -18,12 +18,22 @@ class EnrichedCustomerDataTest extends Setup
 
     public function testToString()
     {
+        $address = [
+            "streetAddress" => "a-street-address",
+            "extendedAddress" => "an-extended-address",
+            "locality" => "a-locality",
+            "region" => "a-region",
+            "postalCode" => "a-code",
+        ];
+
         $profileDataParams = [
             "username" => "venmo_username",
             "firstName" => "John",
             "lastName" => "Doe",
             "phoneNumber" => "1231231234",
             "email" => "john.doe@paypal.com",
+            "billingAddress" => $address,
+            "shippingAddress" => $address,
         ];
 
         $enrichedCustomerDataParams = [
@@ -33,6 +43,6 @@ class EnrichedCustomerDataTest extends Setup
 
         $enrichedCustomerData = Braintree\EnrichedCustomerData::factory($enrichedCustomerDataParams);
 
-        $this->assertEquals("Braintree\EnrichedCustomerData[fieldsUpdated=0=username, profileData=Braintree\VenmoProfileData[username=venmo_username, firstName=John, lastName=Doe, phoneNumber=1231231234, email=john.doe@paypal.com]]", (string) $enrichedCustomerData);
+        $this->assertEquals("Braintree\EnrichedCustomerData[fieldsUpdated=0=username, profileData=Braintree\VenmoProfileData[username=venmo_username, firstName=John, lastName=Doe, phoneNumber=1231231234, email=john.doe@paypal.com, billingAddress=Braintree\Address[streetAddress=a-street-address, extendedAddress=an-extended-address, locality=a-locality, region=a-region, postalCode=a-code], shippingAddress=Braintree\Address[streetAddress=a-street-address, extendedAddress=an-extended-address, locality=a-locality, region=a-region, postalCode=a-code]]]", (string) $enrichedCustomerData);
     }
 }